% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{ApproxCholConfig}
\alias{ApproxCholConfig}
\title{Approximate Cholesky configuration}
\usage{
ApproxCholConfig(seed = 0, split_merge = NULL)
}
\arguments{
\item{seed}{Non-negative integer random seed.}

\item{split_merge}{Optional positive integer split/merge parameter, or \code{NULL}.}
}
\value{
A local-solver configuration object.
}
\description{
Approximate Cholesky configuration
}
