% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{ApproxSchurConfig}
\alias{ApproxSchurConfig}
\title{Approximate Schur complement configuration}
\usage{
ApproxSchurConfig(seed = 0, split = 1L)
}
\arguments{
\item{seed}{Non-negative integer random seed.}

\item{split}{Positive integer edge split factor.}
}
\value{
A local-solver configuration object.
}
\description{
Approximate Schur complement configuration
}
