% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestr.R
\name{closestr}
\alias{closestr}
\title{Closest Location}
\usage{
closestr(longitude, latitude, data)
}
\arguments{
\item{longitude}{numeric}

\item{latitude}{numeric}

\item{data}{data.frame}
}
\value{
index numeric
}
\description{
The goal of withinr package is to find a location index within a certain Radius and a closest location.
You need to rename column 'latitude' to 'lat' and 'longitude' to 'lon'.
}
\examples{
df_sample <- data.frame(lon =  c(128.1, 128, 128.6, 135), lat = c(37, 37.1, 37.5, 42))
closestr(128, 37, df_sample)
}
