% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\docType{data}
\name{ReductionStrategy}
\alias{ReductionStrategy}
\title{Reduction strategy shortcut values}
\format{
An object of class \code{within_reduction_strategy} of length 3.
}
\usage{
ReductionStrategy
}
\description{
Use these values inside \code{\link[=AdditiveSchwarz]{AdditiveSchwarz()}}.
}
\keyword{datasets}
