% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinr.R
\name{withinr}
\alias{withinr}
\title{Location Within a certain Radius}
\usage{
withinr(longitude, latitude, data, dist)
}
\arguments{
\item{longitude}{numeric}

\item{latitude}{numeric}

\item{data}{data.frame}

\item{dist}{numeric}
}
\value{
index numeric
}
\description{
The goal of withinr package is to find a location index within a certain Radius and a closest location.
You need to rename column 'latitude' to 'lat' and 'longitude' to 'lon'.
Distance unit is meter(m).
}
\examples{
df_sample <- data.frame(lon =  c(128.1, 128, 128.6, 135), lat = c(37, 37.1, 37.5, 42))
withinr(128, 37, df_sample, 15000)
}
