% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\docType{data}
\name{PreconditionerConfig}
\alias{PreconditionerConfig}
\title{Preconditioner shortcut values}
\format{
An object of class \code{within_preconditioner_config} of length 3.
}
\usage{
PreconditionerConfig
}
\description{
Use \code{PreconditionerConfig$Additive}, \code{PreconditionerConfig$Off}, or
\code{PreconditionerConfig$Diagonal} as the \code{preconditioner} argument.
}
\keyword{datasets}
